create JuMP extension
